requirements.txt
Django==2.1.2
django-easy-pdf==0.1.1
django-nose==1.4.5
factory-boy==2.11.1
Faker==0.9.1
html5lib==1.0.1
nose==1.3.7
Pillow==5.3.0
PyPDF2==1.26.0
python-dateutil==2.7.3
pytz==2018.5
reportlab==3.5.9
six==1.11.0
text-unidecode==1.2
webencodings==0.5.1
xhtml2pdf==0.2.3